+++ macholib/dyld.py
@@ -4,12 +4,43 @@ dyld emulation
 
 import os
 import sys
+import ctypes
+import platform
 from itertools import chain
 
 from macholib.dylib import dylib_info
 from macholib.framework import framework_info
 
 __all__ = ["dyld_find", "framework_find", "framework_info", "dylib_info"]
+
+if sys.platform == "darwin" and [int(x) for x in platform.mac_ver()[0].split('.')[:2]] >= [10, 16]:
+    try:
+        libc = ctypes.CDLL("libSystem.dylib")
+
+    except OSError: 
+        _dyld_shared_cache_contains_path = None
+
+    else:
+        try:
+            _dyld_shared_cache_contains_path = libc._dyld_shared_cache_contains_path
+        except AttributeError:
+            _dyld_shared_cache_contains_path = None
+
+        else:
+            _dyld_shared_cache_contains_path.restype = ctypes.c_bool
+            _dyld_shared_cache_contains_path.argtypes = [ctypes.c_char_p]
+
+            if sys.version_info[0] != 2:
+                __dyld_shared_cache_contains_path= _dyld_shared_cache_contains_path
+                def _dyld_shared_cache_contains_path(path):
+                    return __dyld_shared_cache_contains_path(path.encode())
+
+    print(libc)
+
+
+
+else:
+    _dyld_shared_cache_contains_path = None
 
 # These are the defaults as per man dyld(1)
 #
@@ -168,6 +199,8 @@ def dyld_find(name, executable_path=None, env=None, loader_path=None):
         ),
         env,
     ):
+        if _dyld_shared_cache_contains_path is not None and _dyld_shared_cache_contains_path(path):
+            return path
         if os.path.isfile(path):
             return path
     raise ValueError("dylib %s could not be found" % (name,))